Gut native messaging and refactor all related functionality
Problem: Tridactyl initializes a _ton_ of state on module load. This
used to be fine because we also used non-typesafe RPCs to obliterate
API boundaries and encapsulation all over the place. Getting rid of
taht and going back to proper use of modules necessarily requires
importing modules that are initializing global state at module load
time. The particular method I've chosen for implementing the proxies
_still loads all of this stuff even if it's not used_, which causes
Badness e.g. the commandline iframe having another commandline iframe
injected into it and the content scripts all running background-script
stuff on load.
