+++ src/background_lib/clipboard.ts
@@ -1,4 +1,4 @@
-import * as Native from "@src/lib/native"
+import Native from "@src/lib/generated/native"
 import { messageActiveTab } from "@src/lib/messaging"
 import { EditorCmds } from "@src/background/editor"
 import * as config from "@src/lib/config"
@@ -18,7 +18,7 @@ async function setclip(str) {
     // Functions to avoid retyping everything everywhere
 
     // Note: We're using fillcmdline here because exceptions are somehow not caught. We're rethrowing because otherwise the error message will be overwritten with the "yank successful" message.
-    const s = () => Native.clipboard("set", str)
+    const s = () => setClipboardContents(str)
     const c = () => messageActiveTab("commandline_frame", "setClipboard", [str])
 
     let promises = []
@@ -41,7 +41,7 @@ export async function getclip(fromm?: "clipboard" | "selection") {
     if (fromm === "clipboard") {
         return messageActiveTab("commandline_frame", "getClipboard")
     } else {
-        return Native.clipboard("get", "")
+        return getClipboardContents()
     }
 }
 
@@ -113,4 +113,87 @@ export async function geturlsforlinks(reltype = "rel", rel: string) {
     const elems = document.querySelectorAll("link[" + reltype + "='" + rel + "']")
     if (elems) return Array.prototype.map.call(elems, x => x.href)
     return []
+}
+
+export async function getClipboardCommand() {
+    const clipcmd = await config.get("externalclipboardcmd")
+    if (clipcmd === "auto") {
+        return Native.firstinpath(["xsel", "xclip"])
+    } else if (clipcmd === undefined) {
+        throw new Error("Couldn't find an external clipboard executable")
+    } else {
+        return clipcmd
+    }
+}
+
+async function setClipboardPost17(contents: string) {
+    // If we have a high enough version, we communicate with the
+    // native messenger using structured messages and can continue
+    // without any escaping mumbo-jumbo.
+    const clipcmd = getClipboardCommand()
+    return Native.run(`${clipcmd} -i`, contents)
+}
+
+async function setClipboardPre17(contents: string) {
+    // Older versions of the native messenger don't support structured
+    // messages, so we have to use a hacky old fashioned way of escaping
+    // the clipboard contents.
+    //
+    // We're going to pretend that we don't know about stdin, and we
+    // need to insert str, which we can't trust, into the clipcmd.
+    //
+    // In order to do this safely we'll use here documents:
+    // http://pubs.opengroup.org/onlinepubs/009695399/utilities/xcu_chap02.html#tag_02_07_04
+    const clipcmd = getClipboardCommand()
+
+    // Find a delimiter that isn't in str
+    let heredoc = "TRIDACTYL"
+    while (contents.search(heredoc) !== -1)
+        heredoc += Math.round(Math.random() * 10)
+
+    // Use delimiter to insert str into clipcmd's stdin. Use sed to
+    // remove the newline added by the here document
+    const invocation = `sed -z 's/.$//' <<'${heredoc}' | ${clipcmd} -i \n${contents}\n${heredoc}`
+    return Native.run(invocation)
+}
+
+/**
+ *
+ * Calls an external program to set the contents of the clipboard. We
+ * have this in addition to the clipboard handling in
+ * commandline_content because the web clipboard APIs can't do
+ * anything with the X selection.
+ *
+ */
+export async function setClipboardContents(contents: string) {
+    let runResult
+    if (await Native.nativegate("0.1.7", false)) {
+        runResult = await setClipboardPost17(contents)
+    } else {
+        runResult = await setClipboardPre17(contents)
+    }
+    if (runResult.code !== 0) {
+        throw new Error(
+            `External command failed with code ${runResult.code}: ${runResult.cmd}`,
+        )
+    }
+}
+
+/**
+ *
+ * Calls an external program to get the contents of the clipboard. We
+ * have this in addition to the clipboard handling in
+ * commandline_content because the web clipboard APIs can't do
+ * anything with the X selection.
+ *
+ */
+export async function getClipboardContents() {
+    const clipcmd = getClipboardCommand()
+    const result = await Native.run(clipcmd + " -o")
+    if (result.code !== 0) {
+        throw new Error(
+            `External command failed with code ${result.code}: ${result.cmd}`,
+        )
+    }
+    return result.content
 }
