+++ src/completions/Preferences.ts
@@ -1,5 +1,5 @@
 import * as Completions from "@src/completions"
-import * as Native from "@src/lib/native"
+import prefs from "@src/lib/generated/about_config"
 
 class PreferenceCompletionOption extends Completions.CompletionOptionHTML
     implements Completions.CompletionOptionFuse {
@@ -39,7 +39,7 @@ export class PreferenceCompletionSource extends Completions.CompletionSourceFuse
             return
         }
         this.lastExstr = exstr
-        const preferences = await Native.getPrefs()
+        const preferences = await prefs.getPrefs()
         this.options = Object.keys(preferences)
             .filter(key => key.startsWith(pref))
             .map(key => new PreferenceCompletionOption(key, preferences[key]))
