+++ src/content.ts
@@ -87,7 +87,7 @@ import * as webext from "@src/lib/webext"
 import Mark from "mark.js"
 import * as perf from "@src/perf"
 import * as keyseq from "@src/lib/keyseq"
-import * as native from "@src/lib/native"
+import * as native from "@src/lib/generated/native"
 import * as styling from "@src/content/styling"
 import { EditorCmds as editor } from "@src/content/editor"
 import * as updates from "@src/lib/updates"
